Renaming to Sling Model Persist, fixed test error with JDK 11 (PostConstruct wasn't imported, needed to add module dependency on javax annotations)
