+++ SlingModelPersist/src/main/assembly/felix.xml
